Refactor *_conver.rb to move common parts to base_convert.rb
