+++ lib/isodoc/sample/pdf_convert.rb
@@ -1,6 +1,5 @@
+require_relative "base_convert"
 require "isodoc"
-require_relative "metadata"
-require "fileutils"
 
 module IsoDoc
   module Sample
@@ -9,12 +8,6 @@ module IsoDoc
     class PdfConvert < IsoDoc::PdfConvert
       def initialize(options)
         @libdir = File.dirname(__FILE__)
-        super
-      end
-
-      def convert1(docxml, filename, dir)
-        FileUtils.cp html_doc_path('logo.jpg'), File.join(@localdir, "logo.jpg")
-        @files_to_delete << File.join(@localdir, "logo.jpg")
         super
       end
 
@@ -33,10 +26,6 @@ module IsoDoc
           htmlintropage: html_doc_path("html_sample_intro.html"),
           scripts_pdf: html_doc_path("scripts.pdf.html"),
         }
-      end
-
-      def metadata_init(lang, script, labels)
-        @meta = Metadata.new(lang, script, labels)
       end
 
       def googlefonts()
@@ -59,51 +48,7 @@ module IsoDoc
         docxml
       end
 
-      def annex_name(annex, name, div)
-        div.h1 **{ class: "Annex" } do |t|
-          t << "#{get_anchors[annex['id']][:label]} "
-          t.br
-          t.b do |b|
-            name&.children&.each { |c2| parse(c2, b) }
-          end
-        end
-      end
-
-      def term_defs_boilerplate(div, source, term, preface)
-        if source.empty? && term.nil?
-          div << @no_terms_boilerplate
-        else
-          div << term_defs_boilerplate_cont(source, term)
-        end
-      end
-
-      def i18n_init(lang, script)
-        super
-        @annex_lbl = "Appendix"
-      end
-
-      def fileloc(loc)
-        File.join(File.dirname(__FILE__), loc)
-      end
-
-      def cleanup(docxml)
-        super
-        term_cleanup(docxml)
-      end
-
-      def term_cleanup(docxml)
-        docxml.xpath("//p[@class = 'Terms']").each do |d|
-          h2 = d.at("./preceding-sibling::*[@class = 'TermNum'][1]")
-          h2.add_child("&nbsp;")
-          h2.add_child(d.remove)
-        end
-        docxml
-      end
-
-      def info(isoxml, out)
-        @meta.security isoxml, out
-        super
-      end
+      include BaseConvert
     end
   end
 end
